Start working on updating rhino gcv plugin
Like the librt code, only more so, these openNURBS API calls have
changed quite a bit.  It looks like the original Audit() call is
completely removed.  It may be the bResolveIdAndNameConflicts flag takes
over this role.  If we do still need to do our own post-processing to
correct things, we'll have to write a new version.

The color and material pieces are also very different - start seeing if
we can find equivalents in the new API.  Will have to be a lot of
checking for this, and may warrant setting up some targeted regression
tests.