Comment out code using removed openNURBS stubs
As near as I can tell, this code was just set up to call methods in
openNURBS that were unimplemented - just stubs that stood in for the
full methods in the RhinoSDK.  Newer openNURBS has removed them all
together, so this no longer compiles in addition to no longer working.
Leaving commented out to document intent for eventual future work that
may want to get this running using functionality in libbrep.